+++ challenges/diceRepresentation/fiveSidedDieSpec.md
